refactor: 테스트 코드 계층화
상속을 통해 어노테이션 재활용
