+++ src/test/java/roomescape/acceptance/PageLanderTest.java
@@ -3,16 +3,9 @@ package roomescape.acceptance;
 import io.restassured.RestAssured;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.transaction.annotation.Transactional;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-@TestPropertySource(locations = "classpath:application-test.properties")
-class PageLanderTest {
+class PageLanderTest extends AcceptanceTest {
 
     @ParameterizedTest(name = "{0} ,{1}")
     @CsvSource(value = {
