+++ src/test/java/roomescape/acceptance/admin/ThemeAcceptanceTest.java
@@ -4,16 +4,12 @@ import io.restassured.RestAssured;
 import io.restassured.common.mapper.TypeRef;
 import io.restassured.http.ContentType;
 import io.restassured.response.ValidatableResponse;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.transaction.annotation.Transactional;
+import roomescape.acceptance.AcceptanceTest;
+import roomescape.acceptance.NestedAcceptanceTest;
 import roomescape.controller.exception.CustomExceptionResponse;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
@@ -27,18 +23,7 @@ import static roomescape.acceptance.Fixture.adminToken;
 import static roomescape.acceptance.PreInsertedData.preInsertedTheme1;
 import static roomescape.acceptance.PreInsertedData.preInsertedTheme2;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-@TestPropertySource(locations = "classpath:application-test.properties")
-class ThemeAcceptanceTest {
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    private void setUp() {
-        RestAssured.port = port;
-    }
+class ThemeAcceptanceTest extends AcceptanceTest {
 
     @DisplayName("관리자가 테마 목록을 조회한다.")
     @Test
@@ -75,7 +60,7 @@ class ThemeAcceptanceTest {
 
     @DisplayName("관리자가 테마를 삭제한다.")
     @Nested
-    class deleteTheme {
+    class deleteTheme extends NestedAcceptanceTest {
 
         @DisplayName("정상 작동")
         @Test
